+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Fixed issues with proxy information not getting properly passed through to app.
   Via [#216](//github.com/threerings/getdown/pull/216).
 
+* `appbase` and `latest` properties in `getdown.txt` now process env var subtitutions.
+
 ## 1.8.4 - May 14, 2019
 
 * Added `verify_timeout` config to allow customization of the default (60 second) timeout during